[fixed] pet deaggro bug. [fixed] faction points bug.
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1478 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/player/PlayerImplementation.cpp
@@ -4702,13 +4702,13 @@ void PlayerImplementation::sendRadialResponseTo(Player* player, ObjectMenuRespon
 }
 
 void PlayerImplementation::addFactionPoints(String faction, uint32 points) {
-	int currentPoints = factionPointsMap.getFactionPoints(faction);
+	uint32 currentPoints = factionPointsMap.getFactionPoints(faction);
 
 	uint32 maxPoints = getMaxFactionPoints(faction);
 	uint32 pointsToAdd = 0;
 
-	if ((currentPoints + (int) points) > maxPoints)
-		pointsToAdd = (int) maxPoints - currentPoints;
+	if ((currentPoints + points) > maxPoints)
+		pointsToAdd = maxPoints - currentPoints;
 	else
 		pointsToAdd = points;
